% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling-locations-model.R
\name{check_duplicate_coordinates}
\alias{check_duplicate_coordinates}
\title{Check for duplicate coordinates}
\usage{
check_duplicate_coordinates(sampling_locations, tolerance = 1e-06)
}
\arguments{
\item{sampling_locations}{Data frame with sampling locations}

\item{tolerance}{Numeric tolerance for coordinate comparison}
}
\value{
List with duplicate checking results
}
\description{
Check for duplicate coordinates
}
