% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization-result-model.R
\name{create_default_metrics}
\alias{create_default_metrics}
\title{Create default performance metrics}
\usage{
create_default_metrics()
}
\value{
Default metrics list
}
\description{
Create default performance metrics
}
