% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-handling.R
\name{create_error}
\alias{create_error}
\title{Create a custom error condition}
\usage{
create_error(message, class, call = sys.call(-1), ...)
}
\arguments{
\item{message}{Error message}

\item{class}{Error class name}

\item{call}{Call expression}
}
\value{
error condition object
}
\description{
Create a custom error condition
}
