% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization-result-model.R
\name{create_ml_result_metadata}
\alias{create_ml_result_metadata}
\title{Create ML result metadata}
\usage{
create_ml_result_metadata(metadata = NULL, method = "unknown")
}
\arguments{
\item{metadata}{Raw metadata}

\item{method}{ML method}
}
\value{
Standardized ML metadata list
}
\description{
Create ML result metadata
}
