% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml-sampling-tool.R
\name{create_ml_sampling_tool}
\alias{create_ml_sampling_tool}
\title{Create Enhanced ML Sampling Tool}
\usage{
create_ml_sampling_tool(config = list(), interactive = FALSE)
}
\arguments{
\item{config}{Configuration list}

\item{interactive}{Whether to run in interactive mode}
}
\value{
MLSampling instance
}
\description{
Create Enhanced ML Sampling Tool
}
