% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress-manager.R
\name{create_progress_manager}
\alias{create_progress_manager}
\title{Create default progress manager}
\usage{
create_progress_manager(config_manager = NULL)
}
\arguments{
\item{config_manager}{Optional config manager}
}
\value{
ProgressManager instance
}
\description{
Create default progress manager
}
