% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource-manager.R
\name{create_resource_manager}
\alias{create_resource_manager}
\title{Create default resource manager}
\usage{
create_resource_manager(config_manager = NULL)
}
\arguments{
\item{config_manager}{Optional config manager}
}
\value{
ResourceManager instance
}
\description{
Create default resource manager
}
