% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization-result-model.R
\name{create_result_metadata}
\alias{create_result_metadata}
\title{Create result metadata}
\usage{
create_result_metadata(metadata = NULL)
}
\arguments{
\item{metadata}{Raw metadata}
}
\value{
Standardized metadata list
}
\description{
Create result metadata
}
