% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml-sampling-tool.R
\name{create_soil_sampling_tool}
\alias{create_soil_sampling_tool}
\title{Create default SoilSamplingTool instance (Deprecated)}
\usage{
create_soil_sampling_tool(...)
}
\arguments{
\item{...}{Additional configuration parameters}
}
\value{
MLSampling instance
}
\description{
This function is deprecated. Please use \code{create_ml_sampling_tool()} instead.
Maintained for backward compatibility.
}
\examples{
\dontrun{
tool <- create_soil_sampling_tool()
}

}
