% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-uncertainty.R
\name{create_spatial_uncertainty}
\alias{create_spatial_uncertainty}
\title{Create Spatial Uncertainty instance}
\usage{
create_spatial_uncertainty(...)
}
\arguments{
\item{...}{Additional configuration parameters}
}
\value{
SpatialUncertainty instance
}
\description{
Convenience function to create a SpatialUncertainty instance with default configuration.
}
\examples{
\dontrun{
uncertainty <- create_spatial_uncertainty()
}

}
