% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design-comparison.R
\name{DesignComparison}
\alias{DesignComparison}
\title{DesignComparison}
\description{
Provides tools for comparing different sampling designs based on coverage,
efficiency, and spatial representativeness.
}
\details{
The DesignComparison class implements:
\itemize{
\item Comparison metrics: Coverage (MSSD), Efficiency, Representativeness
\item Statistical testing: Wilcoxon rank-sum test, t-test
\item Spatial cross-validation for design assessment
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DesignComparison-new}{\code{DesignComparison$new()}}
\item \href{#method-DesignComparison-compare_designs}{\code{DesignComparison$compare_designs()}}
\item \href{#method-DesignComparison-perform_statistical_test}{\code{DesignComparison$perform_statistical_test()}}
\item \href{#method-DesignComparison-cross_validate_design}{\code{DesignComparison$cross_validate_design()}}
\item \href{#method-DesignComparison-clone}{\code{DesignComparison$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignComparison-new"></a>}}
\if{latex}{\out{\hypertarget{method-DesignComparison-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignComparison$new(config = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{Optional configuration list
Compare Multiple Designs}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignComparison-compare_designs"></a>}}
\if{latex}{\out{\hypertarget{method-DesignComparison-compare_designs}{}}}
\subsection{Method \code{compare_designs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignComparison$compare_designs(designs, field_data, true_values = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{designs}}{List of sf objects representing different sampling designs}

\item{\code{field_data}}{Field data (SpatRaster or list with boundary/covariates)}

\item{\code{true_values}}{Optional raster of true values for validation (if available)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List containing comparison results and metrics
Perform Statistical Test Between Two Designs
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignComparison-perform_statistical_test"></a>}}
\if{latex}{\out{\hypertarget{method-DesignComparison-perform_statistical_test}{}}}
\subsection{Method \code{perform_statistical_test()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignComparison$perform_statistical_test(
  design1_metrics,
  design2_metrics,
  test_type = "wilcoxon"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{design1_metrics}}{Vector of metric values for design 1}

\item{\code{design2_metrics}}{Vector of metric values for design 2}

\item{\code{test_type}}{"wilcoxon" or "t-test"}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Test result (htest object)
Spatial Cross-Validation for Design Assessment
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignComparison-cross_validate_design"></a>}}
\if{latex}{\out{\hypertarget{method-DesignComparison-cross_validate_design}{}}}
\subsection{Method \code{cross_validate_design()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignComparison$cross_validate_design(design, field_data, k = 5)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{design}}{sf object of sampling locations}

\item{\code{field_data}}{Field data}

\item{\code{k}}{Number of folds}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
CV performance metrics
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignComparison-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DesignComparison-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignComparison$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
