% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-data-model.R
\name{determine_primary_crs}
\alias{determine_primary_crs}
\title{Determine primary CRS from field data}
\usage{
determine_primary_crs(field_data)
}
\arguments{
\item{field_data}{Field data list}
}
\value{
Primary CRS string
}
\description{
Determine primary CRS from field data
}
