% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization-result-model.R
\name{extract_metric_value}
\alias{extract_metric_value}
\title{Extract metric value safely}
\usage{
extract_metric_value(metrics, metric_name)
}
\arguments{
\item{metrics}{Metrics list}

\item{metric_name}{Name of metric}
}
\value{
Metric value or NA
}
\description{
Extract metric value safely
}
