% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization-result-model.R
\name{extract_ml_component}
\alias{extract_ml_component}
\title{Extract ML component from ml_components list}
\usage{
extract_ml_component(ml_components, component_name, default = NULL)
}
\arguments{
\item{ml_components}{List with ML components}

\item{component_name}{Name of component to extract}

\item{default}{Default value if component not found}
}
\value{
Component value or default
}
\description{
Extract ML component from ml_components list
}
