% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling-locations-model.R
\name{generate_sample_ids}
\alias{generate_sample_ids}
\title{Generate automatic sample IDs}
\usage{
generate_sample_ids(n_locations, types)
}
\arguments{
\item{n_locations}{Number of locations}

\item{types}{Vector of location types}
}
\value{
Character vector of sample IDs
}
\description{
Generate automatic sample IDs
}
