% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization-result-model.R,
%   R/uncertainty-quantification-model.R
\name{\%||\%}
\alias{\%||\%}
\title{Helper function for NULL coalescing}
\usage{
x \%||\% y

x \%||\% y
}
\arguments{
\item{x}{First value}

\item{y}{Second value (if x is NULL)}
}
\value{
Non-NULL value

Non-NULL value
}
\description{
Helper function for NULL coalescing

Helper function for NULL coalescing
}
