% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty-quantification-model.R
\name{standardize_confidence_intervals}
\alias{standardize_confidence_intervals}
\title{Standardize confidence intervals}
\usage{
standardize_confidence_intervals(confidence_intervals)
}
\arguments{
\item{confidence_intervals}{Raw confidence intervals}
}
\value{
Standardized confidence intervals list
}
\description{
Standardize confidence intervals
}
