% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization-result-model.R
\name{standardize_existing_samples}
\alias{standardize_existing_samples}
\title{Standardize existing samples format}
\usage{
standardize_existing_samples(existing_samples)
}
\arguments{
\item{existing_samples}{Raw existing samples data}
}
\value{
Standardized data frame or NULL
}
\description{
Standardize existing samples format
}
