% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization-result-model.R
\name{standardize_location_output}
\alias{standardize_location_output}
\title{Standardize location output format}
\usage{
standardize_location_output(selected_locations)
}
\arguments{
\item{selected_locations}{Raw selected locations}
}
\value{
Standardized data frame
}
\description{
Standardize location output format
}
