% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-data-model.R
\name{standardize_ml_metadata}
\alias{standardize_ml_metadata}
\title{Standardize ML metadata structure}
\usage{
standardize_ml_metadata(ml_metadata, covariates)
}
\arguments{
\item{ml_metadata}{Raw ML metadata}

\item{covariates}{Reference SpatRaster for context}
}
\value{
Standardized ML metadata list
}
\description{
Standardize ML metadata structure
}
