% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization-result-model.R
\name{standardize_ml_performance_metrics}
\alias{standardize_ml_performance_metrics}
\title{Standardize ML performance metrics}
\usage{
standardize_ml_performance_metrics(metrics, method)
}
\arguments{
\item{metrics}{Raw metrics list}

\item{method}{ML method for context}
}
\value{
Standardized ML metrics list
}
\description{
Standardize ML performance metrics
}
