% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization-result-model.R
\name{standardize_optimization_parameters}
\alias{standardize_optimization_parameters}
\title{Standardize optimization parameters}
\usage{
standardize_optimization_parameters(parameters)
}
\arguments{
\item{parameters}{Raw parameters list}
}
\value{
Standardized parameters list
}
\description{
Standardize optimization parameters
}
