% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty-quantification-model.R
\name{standardize_uncertainty_component}
\alias{standardize_uncertainty_component}
\title{Standardize uncertainty component}
\usage{
standardize_uncertainty_component(uncertainty_component, component_type)
}
\arguments{
\item{uncertainty_component}{Raw uncertainty component}

\item{component_type}{Type of uncertainty component}
}
\value{
Standardized uncertainty component
}
\description{
Standardize uncertainty component
}
