% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty-quantification-model.R
\name{standardize_uncertainty_rasters}
\alias{standardize_uncertainty_rasters}
\title{Standardize uncertainty rasters}
\usage{
standardize_uncertainty_rasters(
  uncertainty_rasters,
  epistemic,
  aleatoric,
  total
)
}
\arguments{
\item{uncertainty_rasters}{Raw uncertainty rasters}

\item{epistemic}{Epistemic uncertainty for reference}

\item{aleatoric}{Aleatoric uncertainty for reference}

\item{total}{Total uncertainty for reference}
}
\value{
Standardized uncertainty rasters list
}
\description{
Standardize uncertainty rasters
}
