% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty-quantification-model.R
\name{standardize_uncertainty_validation_metrics}
\alias{standardize_uncertainty_validation_metrics}
\title{Standardize uncertainty validation metrics}
\usage{
standardize_uncertainty_validation_metrics(validation_metrics)
}
\arguments{
\item{validation_metrics}{Raw validation metrics}
}
\value{
Standardized validation metrics list
}
\description{
Standardize uncertainty validation metrics
}
