% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling-locations-model.R
\name{validate_and_expand_vector}
\alias{validate_and_expand_vector}
\title{Validate and expand vector to match required length}
\usage{
validate_and_expand_vector(vector, target_length, vector_name)
}
\arguments{
\item{vector}{Input vector}

\item{target_length}{Required length}

\item{vector_name}{Name for error messages}
}
\value{
Expanded vector
}
\description{
Validate and expand vector to match required length
}
