% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization-result-model.R
\name{validate_compliance_structure}
\alias{validate_compliance_structure}
\title{Validate compliance structure}
\usage{
validate_compliance_structure(compliance_structure)
}
\arguments{
\item{compliance_structure}{Constitutional compliance object}
}
\value{
List with validation results
}
\description{
Validate compliance structure
}
