% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-validation.R
\name{validate_data_quality}
\alias{validate_data_quality}
\title{Validate data quality standards}
\usage{
validate_data_quality(covariates, strict_validation = TRUE)
}
\arguments{
\item{covariates}{SpatRaster with environmental data}

\item{strict_validation}{Logical for constitutional standards}
}
\value{
List with data quality results
}
\description{
Validate data quality standards
}
