% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-data-model.R
\name{validate_field_boundary_geometry}
\alias{validate_field_boundary_geometry}
\title{Validate boundary geometry}
\usage{
validate_field_boundary_geometry(boundary, strict_validation = TRUE)
}
\arguments{
\item{boundary}{sf object with field boundary}

\item{strict_validation}{Logical for strict validation mode}
}
\value{
List with validation results
}
\description{
Validate boundary geometry
}
