% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling-locations-model.R
\name{validate_location_types}
\alias{validate_location_types}
\title{Validate location types}
\usage{
validate_location_types(types)
}
\arguments{
\item{types}{Character vector of location types}
}
\value{
List with validation results
}
\description{
Validate location types
}
