% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty-quantification-model.R
\name{validate_mc_samples}
\alias{validate_mc_samples}
\title{Validate Monte Carlo samples}
\usage{
validate_mc_samples(mc_samples)
}
\arguments{
\item{mc_samples}{Monte Carlo samples}
}
\value{
Validated MC samples or NULL
}
\description{
Validate Monte Carlo samples
}
