% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty-quantification-model.R
\name{validate_mc_samples_consistency}
\alias{validate_mc_samples_consistency}
\title{Validate Monte Carlo samples consistency}
\usage{
validate_mc_samples_consistency(mc_samples, n_samples)
}
\arguments{
\item{mc_samples}{Monte Carlo samples}

\item{n_samples}{Number of samples}
}
\value{
List with validation results
}
\description{
Validate Monte Carlo samples consistency
}
