% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization-result-model.R
\name{validate_ml_method}
\alias{validate_ml_method}
\title{Validate ML method specification}
\usage{
validate_ml_method(method)
}
\arguments{
\item{method}{Character string with ML method}
}
\value{
Validated method string
}
\description{
Validate ML method specification
}
