% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty-quantification-model.R
\name{validate_n_samples}
\alias{validate_n_samples}
\title{Validate number of samples}
\usage{
validate_n_samples(n_samples, mc_samples)
}
\arguments{
\item{n_samples}{Number of samples}

\item{mc_samples}{MC samples for consistency check}
}
\value{
Validated number of samples
}
\description{
Validate number of samples
}
