% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling-locations-model.R
\name{validate_numeric_coordinates}
\alias{validate_numeric_coordinates}
\title{Validate numeric coordinates}
\usage{
validate_numeric_coordinates(sampling_locations)
}
\arguments{
\item{sampling_locations}{Data frame with sampling locations}
}
\value{
List with validation results
}
\description{
Validate numeric coordinates
}
