% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization-result-model.R
\name{validate_result_metrics}
\alias{validate_result_metrics}
\title{Validate result metrics structure}
\usage{
validate_result_metrics(metrics)
}
\arguments{
\item{metrics}{Metrics object}
}
\value{
List with validation results
}
\description{
Validate result metrics structure
}
