% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling-locations-model.R
\name{validate_sample_ids}
\alias{validate_sample_ids}
\title{Validate sample IDs}
\usage{
validate_sample_ids(sample_ids)
}
\arguments{
\item{sample_ids}{Character vector of sample IDs}
}
\value{
List with validation results
}
\description{
Validate sample IDs
}
