% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty-quantification-model.R
\name{validate_single_uncertainty_component}
\alias{validate_single_uncertainty_component}
\title{Validate single uncertainty component}
\usage{
validate_single_uncertainty_component(uncertainty_component, component_name)
}
\arguments{
\item{uncertainty_component}{Single uncertainty estimate}

\item{component_name}{Name for error messages}
}
\value{
List with validation results
}
\description{
Validate single uncertainty component
}
