% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-validation.R
\name{validate_spatial_extent_alignment}
\alias{validate_spatial_extent_alignment}
\title{Validate spatial extent alignment}
\usage{
validate_spatial_extent_alignment(boundary, covariates)
}
\arguments{
\item{boundary}{sf boundary object}

\item{covariates}{SpatRaster covariates object}
}
\value{
List with extent alignment results
}
\description{
Validate spatial extent alignment
}
