% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-data-model.R
\name{validate_spatial_weights}
\alias{validate_spatial_weights}
\title{Validate spatial weights matrix}
\usage{
validate_spatial_weights(spatial_weights)
}
\arguments{
\item{spatial_weights}{Matrix or list of spatial weights}
}
\value{
List with validation results
}
\description{
Validate spatial weights matrix
}
