% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty-quantification-model.R
\name{validate_uncertainty_method}
\alias{validate_uncertainty_method}
\title{Validate uncertainty method}
\usage{
validate_uncertainty_method(method)
}
\arguments{
\item{method}{ML method string}
}
\value{
Validated method string
}
\description{
Validate uncertainty method
}
