% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml-sampling-tool.R
\name{SoilSamplingTool}
\alias{SoilSamplingTool}
\title{Legacy SoilSamplingTool Class (Deprecated)}
\description{
Alias for MLSampling for backward compatibility.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MLSampling-new}{\code{SoilSamplingTool$new()}}
\item \href{#method-MLSampling-run_udl}{\code{SoilSamplingTool$run_udl()}}
\item \href{#method-MLSampling-run_ufn}{\code{SoilSamplingTool$run_ufn()}}
\item \href{#method-MLSampling-run_bdl}{\code{SoilSamplingTool$run_bdl()}}
\item \href{#method-MLSampling-run_rf_optimization}{\code{SoilSamplingTool$run_rf_optimization()}}
\item \href{#method-MLSampling-run_ensemble}{\code{SoilSamplingTool$run_ensemble()}}
\item \href{#method-MLSampling-compare_designs}{\code{SoilSamplingTool$compare_designs()}}
\item \href{#method-MLSampling-compare_models}{\code{SoilSamplingTool$compare_models()}}
\item \href{#method-MLSampling-quantify_uncertainty}{\code{SoilSamplingTool$quantify_uncertainty()}}
\item \href{#method-MLSampling-generate_ml_report}{\code{SoilSamplingTool$generate_ml_report()}}
\item \href{#method-MLSampling-generate_report}{\code{SoilSamplingTool$generate_report()}}
\item \href{#method-MLSampling-save_coordinates_to_csv}{\code{SoilSamplingTool$save_coordinates_to_csv()}}
\item \href{#method-MLSampling-get_supported_algorithms}{\code{SoilSamplingTool$get_supported_algorithms()}}
\item \href{#method-MLSampling-validate_report}{\code{SoilSamplingTool$validate_report()}}
\item \href{#method-MLSampling-clone}{\code{SoilSamplingTool$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLSampling-new"></a>}}
\if{latex}{\out{\hypertarget{method-MLSampling-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SoilSamplingTool$new(
  config = NULL,
  config_manager = NULL,
  validate_system = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLSampling-run_udl"></a>}}
\if{latex}{\out{\hypertarget{method-MLSampling-run_udl}{}}}
\subsection{Method \code{run_udl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SoilSamplingTool$run_udl(
  field_data = NULL,
  existing_samples = NULL,
  n_new_samples,
  optimization_method = "greedy",
  model_config = NULL,
  parallel = FALSE,
  max_iter = NULL,
  save_csv = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLSampling-run_ufn"></a>}}
\if{latex}{\out{\hypertarget{method-MLSampling-run_ufn}{}}}
\subsection{Method \code{run_ufn()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SoilSamplingTool$run_ufn(
  field_data = NULL,
  existing_samples = NULL,
  n_new_samples,
  model_config = NULL,
  force_neural_network = FALSE,
  force_statistical_fallback = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLSampling-run_bdl"></a>}}
\if{latex}{\out{\hypertarget{method-MLSampling-run_bdl}{}}}
\subsection{Method \code{run_bdl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SoilSamplingTool$run_bdl(
  field_data,
  existing_samples,
  n_new_samples,
  uncertainty_type = "total",
  mc_iterations = 100,
  constitutional_compliance = TRUE,
  save_csv = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLSampling-run_rf_optimization"></a>}}
\if{latex}{\out{\hypertarget{method-MLSampling-run_rf_optimization}{}}}
\subsection{Method \code{run_rf_optimization()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SoilSamplingTool$run_rf_optimization(
  field_data,
  existing_samples,
  n_new_samples,
  feature_importance_method = "permutation",
  spatial_autocorr = TRUE,
  constitutional_compliance = TRUE,
  save_csv = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLSampling-run_ensemble"></a>}}
\if{latex}{\out{\hypertarget{method-MLSampling-run_ensemble}{}}}
\subsection{Method \code{run_ensemble()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SoilSamplingTool$run_ensemble(
  field_data,
  existing_samples,
  n_new_samples,
  methods = c("BDL", "RF"),
  ensemble_method = "voting",
  constitutional_compliance = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLSampling-compare_designs"></a>}}
\if{latex}{\out{\hypertarget{method-MLSampling-compare_designs}{}}}
\subsection{Method \code{compare_designs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SoilSamplingTool$compare_designs(
  field_data,
  existing_samples,
  n_new_samples,
  methods = c("BDL", "RF"),
  comparison_metrics = c("coverage", "representativeness"),
  constitutional_compliance = TRUE,
  statistical_test = "wilcoxon",
  detailed_metrics = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLSampling-compare_models"></a>}}
\if{latex}{\out{\hypertarget{method-MLSampling-compare_models}{}}}
\subsection{Method \code{compare_models()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SoilSamplingTool$compare_models(
  field_data = NULL,
  existing_samples = NULL,
  n_new_samples,
  algorithms = self$supported_algorithms[1:3],
  n_iterations = 5,
  parallel = FALSE,
  confidence_level = 0.95,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLSampling-quantify_uncertainty"></a>}}
\if{latex}{\out{\hypertarget{method-MLSampling-quantify_uncertainty}{}}}
\subsection{Method \code{quantify_uncertainty()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SoilSamplingTool$quantify_uncertainty(
  predictions,
  method = "ensemble",
  uncertainty_type = "total"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLSampling-generate_ml_report"></a>}}
\if{latex}{\out{\hypertarget{method-MLSampling-generate_ml_report}{}}}
\subsection{Method \code{generate_ml_report()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SoilSamplingTool$generate_ml_report(
  result,
  report_type = "comprehensive",
  include_uncertainty_analysis = TRUE,
  include_visualizations = TRUE,
  constitutional_compliance = TRUE,
  output_dir = getwd()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLSampling-generate_report"></a>}}
\if{latex}{\out{\hypertarget{method-MLSampling-generate_report}{}}}
\subsection{Method \code{generate_report()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SoilSamplingTool$generate_report(
  optimization_result = NULL,
  output_format = "html",
  report_config = NULL,
  export_path = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLSampling-save_coordinates_to_csv"></a>}}
\if{latex}{\out{\hypertarget{method-MLSampling-save_coordinates_to_csv}{}}}
\subsection{Method \code{save_coordinates_to_csv()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SoilSamplingTool$save_coordinates_to_csv(
  optimization_result = NULL,
  file_path,
  output_crs = NULL,
  include_metadata = FALSE,
  include_crs_info = FALSE,
  decimal_places = 6,
  coordinate_format = "decimal",
  column_names = NULL,
  include_fields = NULL,
  include_covariate_values = FALSE,
  validate_export = TRUE,
  constitutional_compliance = FALSE,
  quality_assurance = FALSE,
  standard_format = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLSampling-get_supported_algorithms"></a>}}
\if{latex}{\out{\hypertarget{method-MLSampling-get_supported_algorithms}{}}}
\subsection{Method \code{get_supported_algorithms()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SoilSamplingTool$get_supported_algorithms()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLSampling-validate_report"></a>}}
\if{latex}{\out{\hypertarget{method-MLSampling-validate_report}{}}}
\subsection{Method \code{validate_report()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SoilSamplingTool$validate_report(report)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLSampling-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MLSampling-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SoilSamplingTool$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
